Defaulting the value of PORTAGE_USERNAME to prevent an exception
Need to retrieve the /etc/passwd defined home directory for the user that
portage will be running under. Defaulting to "portage", of course. This
apparently isn't done elsewhere in the code. I was a little supprised that
that key doesn't appear in settings unless it's explicitly set in make.conf.