Integrated debug chatter into the logging system...
...to the extent that is feasible (Portage's logging infrastructure is complex
and nuanced) so that I can watch for corner cases in the logs. Turns out there's
a real chicken & egg issue with getting log messages in the package build log
when you're doing it from the prepare_build_dirs() function that sets up the
build directories, on which the logging depends.

The compromise approach is to try to log to both the build log and the the
emerge.log file (which does not have these dependencies). That way I have the
messages in the context of the build--where I can--and yet do not lose any
messages where I can't. None of this is going into the actual implementation so,
although it's ugly, I don't care as it's just for my benefit.