New approach to indentation recursion
The __better_repr__ method is *not* responsible for indenting the initial type
name that it prints. This is always handled by the calling function. This works
at the phantom indent level 0 of the initial invocation of __better_repr__
because it *needs* no indentation. But at all other levels of recursion, the
*calling* instance handles the indent of the variable name and then calls the
*next* level of __better_repr__, which prints the type (without further indent
needed because it's part of the already indented line). That took a while to
figure out.

Also changed default max depth from 4 to 16. Should make this part of settings.