Move the __better_repr__ implementation into its module
Since the three initial __better_repr__ implementations ended up being
identical, I made the decision to move the implementation into the better_repr
module. All that's left in each class is a simple wrapper method that calls the
module function, which will dramatically simplify the roll-out to other classes.

If custom display methods become desirable, I will evolve some sort of
registration mechanism so that the bulk of the dump code can *remain* in the
module.