Change first argument to all BetterRepr class methods to "context".
This changes makes it more clear what is being passed around. Client classes use
the conventional "self", whereas the BetterRepr service class uses "context".

Also change the the _dump_depgraph() method to use the BetterRepr._print method
so that *all* lines in the dump file will be accounted for in the line number
tracking.