In _dump_depgraph(), manually printing the first line number...
...to the log before the first call into the __better_repr__ tree. This is
because, for all *recursive* calls deeper into the tree, it's the *calling*
instance that's responsible for printing the line number. Addressing this within
the class appears to require an extra argument to *all* the little __better__repr__
wrapper functions and, since the line number print was created *just* to
validate that the log line number tracking mechanism was working properly, isn't
worth the work.

Also corrected the owner of the log file, which ended up being owned by root when
all other files in PORTAGE_LOGDIR are owned by portage.

Long term, I *think* the log file creation logic needs to become part of the
BetterRepr class. This will remove *most* if the imports I've had to add to
depgraph.py and potentially other dumps I undertake that *aren't* part of the
dependency graph.